drawer segregated, questions are list, questions can be asked
