+++ app/src/main/res/layout/activity_drawer.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <android.support.v4.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/home_drawer_layout"
+    android:id="@+id/drawer_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
@@ -16,19 +16,11 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"/>
 
-        <ProgressBar
-            android:id="@+id/home_progress_spinner"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:indeterminate="true"
-            android:layout_centerInParent="true"/>
-
-        <android.support.v7.widget.RecyclerView
-            android:id="@+id/home_recycler_view"
+        <FrameLayout
+            android:id="@+id/drawer_subclass"
             android:layout_below="@+id/home_toolbar"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="5dp"/>
+            android:layout_height="match_parent"/>
 
     </RelativeLayout>
